update: move mainState to different folder, move adapter to view
