+++ app/src/main/java/com/github/ainul/twisdev/ui/main/view/adapter/listener/ListItemListener.kt
@@ -1,4 +1,4 @@
-package com.github.ainul.twisdev.ui.main.adapter.listener
+package com.github.ainul.twisdev.ui.main.view.adapter.listener
 
 import com.github.ainul.twisdev.ui.main.viewmodel.MainViewModel.Companion.CartItems
