+++ app/src/main/java/com/github/ainul/twisdev/ui/main/view/fragment/MainFragment.kt
@@ -8,12 +8,12 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
 import androidx.navigation.fragment.findNavController
 import com.github.ainul.twisdev.R
-import com.github.ainul.twisdev.ui.main.adapter.GridItemAdapter
-import com.github.ainul.twisdev.ui.main.adapter.listener.GridItemListener
+import com.github.ainul.twisdev.ui.main.view.adapter.GridItemAdapter
+import com.github.ainul.twisdev.ui.main.view.adapter.listener.GridItemListener
 import com.github.ainul.twisdev.databinding.FragmentMainBinding
 import com.github.ainul.twisdev.data.model.ItemModel
 import com.github.ainul.twisdev.ui.main.viewmodel.MainViewModel
-import com.github.ainul.twisdev.ui.main.viewmodel.ViewState
+import com.github.ainul.twisdev.ui.main.viewstate.MainState
 import com.github.ainul.twisdev.util.hide
 import com.github.ainul.twisdev.util.show
 import com.google.android.material.transition.MaterialSharedAxis
@@ -35,9 +35,7 @@ class MainFragment : Fragment(), GridItemListener {
         // set option to begin work with menu & set transition to this screen
         setHasOptionsMenu(true)
         setTransitionAnimation()
-
         setupRecyclerView()
-
         // liveData watcher
         updateLiveData()
 
@@ -68,12 +66,12 @@ class MainFragment : Fragment(), GridItemListener {
     private fun updateLiveData() {
         viewmodel.fetchedListItemData.observe(viewLifecycleOwner, { state ->
             when (state) {
-                is ViewState.Loading -> {
+                is MainState.Loading -> {
                     binding.loader.show()
                     binding.errorMessageContainer.hide()
                 }
 
-                is ViewState.Failure -> {
+                is MainState.Failure -> {
                     binding.loader.hide()
                     binding.errorMessageContainer.show()
                     viewmodel.hideActionBar(true)
@@ -82,7 +80,7 @@ class MainFragment : Fragment(), GridItemListener {
                     // viewmodel.refresh()
                 }
 
-                is ViewState.Succeed -> {
+                is MainState.Succeed -> {
                     binding.loader.hide()
                     binding.errorMessageContainer.hide()
                     viewmodel.hideActionBar(false)
