+++ app/src/main/java/com/github/ainul/twisdev/ui/main/view/fragment/ShoppingCartFragment.kt
@@ -7,8 +7,8 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
 import com.github.ainul.twisdev.R
-import com.github.ainul.twisdev.ui.main.adapter.ListItemAdapter
-import com.github.ainul.twisdev.ui.main.adapter.listener.ListItemListener
+import com.github.ainul.twisdev.ui.main.view.adapter.ListItemAdapter
+import com.github.ainul.twisdev.ui.main.view.adapter.listener.ListItemListener
 import com.github.ainul.twisdev.databinding.FragmentShoppingCartBinding
 import com.github.ainul.twisdev.ui.main.viewmodel.MainViewModel
 import com.google.android.material.transition.MaterialSharedAxis
