+++ app/src/main/java/com/github/ainul/twisdev/ui/main/viewstate/MainState.kt
@@ -1,0 +1,9 @@
+package com.github.ainul.twisdev.ui.main.viewstate
+
+import com.github.ainul.twisdev.data.model.ItemModel
+
+sealed class MainState {
+    object Loading : MainState()
+    data class Succeed(val data: List<ItemModel>) : MainState()
+    data class Failure(val message: String?) : MainState()
+}
